junk/feedprism-poc/pyproject.toml
[project]
name = "feedprism-poc"
version = "0.1.0"
description = "A proof-of-concept for FeedPrism"
dependencies = [
    "openai>=1.50.0",
    "qdrant-client==1.7.0",
    "google-auth-oauthlib==1.1.0",
    "google-auth-httplib2==0.1.1",
    "google-api-python-client==2.108.0",
    "beautifulsoup4==4.12.2",
]
requires-python = ">=3.9"
license = "MIT"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"
